% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.matrix.df.r
\name{fill.matrix.df}
\alias{fill.matrix.df}
\title{Create dispersal or correlation matrix for RAMAS GIS model}
\usage{
fill.matrix.df(PopDF, Func, Type)
}
\arguments{
\item{PopDF}{Population information from the mp.file structure. Specifically
the information in the mp.file$PopList element}

\item{Func}{Dispersal- or correlation-distance function parameters. Element
mp.file$DispDistFunc or mp.file$CorrDistFunc in the mp.file 'list' structure}

\item{Type}{Either "disp" or "corr" designating the type of matrix to create}
}
\value{
Either a dispersal (\code{DispMatr}) or correaltion (\code{CorrMatr}) matrix,
depending on \code{Type} setting
}
\description{
Creates a Dispersal or Correlation Matrix based on the
dispersal-(correlation-)distance function
}
