% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdhist.read.r
\name{hdhist.read}
\alias{hdhist.read}
\title{Read HabDyn history TXT file}
\usage{
hdhist.read(hdhistFile)
}
\arguments{
\item{hdhistFile}{name of HIST file}
}
\value{
A list object containing all for the information in the HIST file
}
\description{
\code{hdhist.read}
reads data from habdyn history TXT file (hd-hist), produced by the RAMAS GIS -
Habitat Dynamics module.  The history (HIST) file provides a key
for determining which populaitons in individual *.ptc files match up
with populations in the *.mp file that is produced when several *.ptc
files are combined by the Habitat Dyanimcs (HabDyn) module.
}
