% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metapopversion.r
\name{metapopversion}
\alias{metapopversion}
\title{Determine RAMAS Metapop version used to create a *.MP file}
\usage{
metapopversion(mpFile)
}
\arguments{
\item{mpFile}{A list object of all of the parameters set in a *.MP file}
}
\value{
RAMAS Metapop version number as an integer
}
\description{
\code{metapopversion} is a function called from within the
\code{mp.read} function. It's single purpose is to read the `list` object
that contains all of the *.MP file values and determine which version
of RAMAS Metapop was used to construct the file.
}

