% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.mult.results.r
\name{mp.mult.results}
\alias{mp.mult.results}
\title{Extract results for a list of *.MP files}
\usage{
mp.mult.results(mp.file.list, spatial = FALSE)
}
\arguments{
\item{mp.file.list}{A list (character vector) of *.MP files to extract
results from}

\item{spatial}{(TRUE/FALSE) Do the *.MP files contain spatial information
to be extracted}
}
\description{
Call \code{mp.results} for a list of files and export results as a csv file
}
\seealso{
\code{\link{mp.results}} for a full list of descriptions for each
of the columns in the resulting CSV file.
}
