% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mp.read.r
\name{mp.read}
\alias{mp.read}
\title{Read RAMAS (c) *.MP file}
\usage{
mp.read(mpFile)
}
\arguments{
\item{mpFile}{The name of the *.MP file to be read.}
}
\value{
\code{mp.read} returns a nested list object. The first level includes
two elements, the version of the *.MP file and list names "mp.file". The
"mp.file" list is composed of 52 elements that provide all of the information
found in the *.MP file.
}
\description{
\code{mp.read} reads in the *.MP file used in the RAMAS Metapop software.
The *.MP file is a simple text file containing all of the values necessary
to configure a metapopulation simulation.
}
\details{
mp.file 'list' structure elements. Most element names from RAMAS Metapop source code
The assignment of each element of the mp.file list is based on the Metapopulation input file
type for version 5.0.
\enumerate{
 \item MaxRep: maximum number of replications (integer; 1-10000)
 \item MaxDur: Duration of simulation (integer; 1-500)
 \item Demog_Stoch: Use demographic stochasticity? (boolean)
 \item Stages: Number of stages (integer; 1-50)
 \item Stages.IgnoreConstraints: Ingore constraints? (boolean - TRUE/FALSE)
 \item Cat1: Information associated with catastrophe 1
   This information is subdivided in RAMAS Input, but will not be subdivided
   in the first version of the sensitivity analysis package
 \item Cat2: Information associated with catastrophe 2
   See note associated with Cat1
 \item DDActing: Information regarding density dependence
 \item Distrib: Distribution (Normal or Lognormal) to use for Environmental
   Stochasticity.
 \item AdvancedStochSetting: Advanced Environmental Stoch. Settings: (0 or PoolVar or
   NegSurvCorr)
 \item dispCV: Coefficient of variation for dispersal
 \item WhenBelow: When below local threshold (count in total or don't count or assume dead)
 \item corrwith: Within population correlation -
   1 - (F, S, K Correlated)
   2 - (F and S correlated, K independent)
   3 - (F, S, K independent
 \item DispersalDependsOnTargetPopK: Just like the variable name implie; boolean
 \item DDBasis: Density basis type (char)
 \item PopSpecificDD: Density dependence type is population specific (boolean - Yes/No)
 \item DDforAllPop: Density dependence type for all populations (char)
 \item UserDllFileName: Filename for user-defined density dependence function
 \item TimeStepSize: Time step size
 \item TimeStepUnits: Time step units
 \item SexStructure: Description of Sex Structure (char)
 \item FemaleStages: Number of female stages
 \item MatingSystem: Description of Mating system selected (char)
 \item FemalesPerMale: Number of females per male as given by user
 \item MalesPerFemale: Number of males per female as given by user
 \item CVerror: sampling error for N
 \item FirstTforRisk: Initial number of time steps to exclude from risk calculations
 Population Parameters
 \item PopList: Individual population data - saved as a 'nested list' structure, one 'list' for
   each population.
  Dispersal (Migration) Parameters
 \item UseDispDistFunc: True if dispersal rates are based on dispersal distance function; false if
    they are specified in the dispersal matrix
 \item DispDistFunc: Dispersal-distance function parameters - a, b, c, Dmax - Mij = a exp(-Dij^c/b)
 \item DispMatr: Dispersal matrix, as defined or calculated in 'fill.matrix.df.r'
 Correlation Parameters
 \item UseCorrDistFunc: True if correlations between populations is based on correlation distance
    function; False if they are specified in the correlation matrix
 \item CorrDistFunc: Correlation-distance function parameters - a, b, c - Cij = a exp(-Dij^c/b)
 \item CorrMatr: User specified correlation matrix, as defined or calculated in 'fill.matrix.df.r'
 Stage Matrices Information
 \item StMatrNumber: number of stage matrices defined by user in *.mp file
 \item StMatr: a 'list' object containing information about each stage matrix.  For each stage
   matrix the following information is stored:
   StMatrName: Name of the stage matrix
   StMatrSurvMult: Survival multiplier for stage matrix (for generating new matrices; not used in simulation)
   StMatrFecMult: Fecundity multiplier for stage matrix (for generating new matrices; not used in simulation)
   Matr: a matrix object with numeric values
 Standard Deviation Matrices Information
 \item SDMatrNumber: number of st. dev. matrices defined by user in *.mp file
 \item SDMatr: a 'list' object containing information for each st.dev. matrix.  For each st.dev.
   matrix the following information is stored:
   SDMatrName: Name of the st.dev. matrix
   Matr: a matrix object of numeric values
 \item ConstraintsMatr: Constraints Matrix
 \item StMig: Stage Relative Migration (Dispersal) Rates
 \item Cat1EffMat: Catastrophe 1 effects on vital rates; one row per stage, seperated by spaces
 \item Cat1EffNst: Catastrophe 1 effects on abundances
 \item Cat2EffMat: Catastrophe 2 effects on vital rates; one row per stage, seperated by spaces
 \item Cat2EffNst: Catastrophe 2 effects on abundances
 \item StInit: Initial Abundance for each stage for each population
  Stage Properties
 \item StProp: a 'list' object containing the following information for each stage
    StName: Name of stage
    StWeight: Relative weight of stage
   StExclude: Exclude from total (boolean: TRUE or FALSE)
   StBasisForDD: Basis for density-dependence (boolean: TRUE or FALSE)
   StBreeding: Proportion of stage physically capable of breeding
  Population Management Actions
 \item NPopManage: Number of population managment actions
 \item PopManageProp: All of the lines associated with population management, unpartitioned
 \item ExtinctThr:  Extinction threshold
 \item ExplodeThr:  Explosion threshold
 \item stepsize: step size
 \item PopData_df: Population level information in data.frame format
}
}

