% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mp.results.r
\name{mp.results}
\alias{mp.results}
\title{Calculates various endpoints using results from RAMAS MP simulation}
\usage{
mp.results(mpFile, spatial = FALSE, ptc = FALSE, ptcFiles = "no file",
  ptcFileIter = "", habdyn = FALSE, hdhFile = "no file")
}
\arguments{
\item{spatial}{TRUE/FALSE - should spatial parameters be extracted from the *.mp file,
including factors such as dispersal, correlation, and factors included in
*.ptc files, if included}

\item{ptc}{TRUE/FALSE - Is a *.ptc file included?  If yes, some spatial information can
be extracted from this file}

\item{ptcFiles}{A single path to a ptc file or a vector of ptc files with which to extract
data from.}

\item{ptcFileIter}{For each ptcFile, the corresponding iteration as noted in the HabDyn
history file.  If no history file is used, iter values can be some user defined
values, however an error will be generated if the length of ptcfiles and
ptcFileIter do not match.}

\item{habdyn}{TRUE/FALSE - Was HabDyn module used? Is there a corresponding HabDyn
history file?}

\item{hdhFile}{The full path name to the HabDyn history (.txt) file}

\item{mpList}{The full list object of an MP file that has been run (i.e., has results)}
}
\value{
res.summ.df A data frame of the endpoints calculated
}
\description{
\code{mp.results} Calculates various endpoints from the results of an MP
simulation as carried out in RAMAS. Use for Metapop version 5 and 5.1
file formats. Based on ResOut.exe program written by R. Akcakaya
}

