% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mp.write.r
\name{mp.write}
\alias{mp.write}
\title{Write new RAMAS (c) *.MP file}
\usage{
mp.write(mp.new, version, mp.new.file)
}
\arguments{
\item{mp.new}{mp.file list structure, the result of the mp.read.r or sensitivity.r functions}

\item{version}{the Metapop version number of the mp.file}

\item{mp.new.file}{file name for the new *.mp file to be written}
}
\value{
A *.MP file will be written. While this file is in simple text
format, it is necessary for RAMAS Metapop simulations.
}
\description{
The input mp.new structure should be the structure resulting from
either the mp.read.r or the sensitivity.r scripts.  This script produces a file
readable by the RAMAS Metapop program.
}

