% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nix2win.r
\name{nix2win}
\alias{nix2win}
\title{Convert MP files from *nix to Windows format}
\usage{
nix2win(mp.new.file)
}
\arguments{
\item{mp.new.file}{Name of the *.mp file to convert from *nix to Windows
format}
}
\value{
There are no returns to the R environment. However, the selected *.mp file
will be in Windows format after the function has been executed.
}
\description{
\code{nix2win} is a wrapper function for a bash script to convert from
*nix formated text files to Windows formated text files. These files have
different end-of-line characters. RAMAS GIS only reads Windows formated
files.
}

