% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptc.read.r
\name{ptc.read}
\alias{ptc.read}
\title{Read RAMAS GIS Patch (PTC) file}
\usage{
ptc.read(ptcFile)
}
\arguments{
\item{ptcFile}{The name of the *.ptc file to read}
}
\value{
\code{ptc.read} returns a list object containing all information
from the *.ptc file.
}
\description{
\code{ptc.read} reads data from a *.ptc file, used in RAMAS GIS -
Spatial Data module. This file contains information on the spatial
structure of the metapopulation.
}

