% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sensitivity.r
\name{sensitivity}
\alias{sensitivity}
\title{RAMAS Metepopulation sensitivity analysis program}
\usage{
sensitivity(sens.config.file)
}
\arguments{
\item{sens.config.file}{A simple text file containing all the necessary
information to run a \code{demgsa} sensitivity analysis.}
}
\value{
There are no direct returns to the R environment. However, new
*.MP files are created in a directory determined by values set in the
sens.config.file.
}
\description{
\code{sensitivity} is the main function for the \code{demgsa} package.
This function compares the input RAMAS *.MP files and setups up the
creation of new *.MP files to run to examine the sensitivity of
model outputs on model input parameter uncertainty
}

